server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.1
attrs==23.2.0
beautifulsoup4==4.10.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.1
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
frozenlist==1.4.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.331
langsmith==0.0.60
marshmallow==3.20.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.23.2
packaging==24.0
Pillow==9.5.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-decouple==3.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
yarl==1.9.4
